Move the test, it needs that importorskip.
