+++ dask/dataframe/io/tests/test_parquet.py
@@ -4517,43 +4517,3 @@ def test_select_filtered_column(tmp_path, engine):
     with pytest.warns(UserWarning, match="Sorted columns detected"):
         ddf = dd.read_parquet(path, engine=engine, filters=[("b", "==", "cat")])
     assert_eq(df, ddf)
-
-
-@PYARROW_MARK
-@pytest.mark.parametrize(
-    "engine,setting,expected_type",
-    [
-        pytest.param(
-            "pyarrow",
-            True,
-            pd.StringDtype(storage="pyarrow"),
-            marks=pytest.mark.skipif(not PANDAS_GT_130, reason="needs pandas>=1.3.0"),
-        ),
-        ("pyarrow", False, object),
-        ("fastparquet", True, object),
-        ("fastparquet", False, object),
-    ],
-)
-def test_read_parquet_convert_string(tmpdir, engine, setting, expected_type):
-    """Test that string dtypes are converted with dd.read_parquet and
-    dataframe.convert_string=True"""
-
-    df = pd.DataFrame({"A": ["def", "abc", "ghi"]}, dtype=object)
-    path = str(tmpdir.join("test.parquet"))
-    df.to_parquet(path)
-
-    ctx = contextlib.nullcontext()
-    success = True
-    if engine == "fastparquet" and setting:
-        ctx = pytest.raises(ValueError, match="`convert_strings` is not supported")
-        success = False
-
-    with dask.config.set({"dataframe.convert_string": setting}):
-        with ctx:
-            ddf = dd.read_parquet(path, engine=engine)
-        if success:
-            assert expected_type == ddf.A.dtype
-            assert ddf.A.dtype == ddf.compute().A.dtype
-            # make sure we didn't convert types in __init__, the only operation
-            # was read_parquet
-            assert len(ddf.dask.layers) == 1
